Move search config to its own file
